--- a/commands/benchmark.go
+++ b/commands/benchmark.go
@@ -57,7 +57,7 @@ func benchmark(cmd *cobra.Command, args []string) error {
 			return err
 		}
 		for i := 0; i < benchmarkTimes; i++ {
-			MainSite = nil
+			MainSites = nil
 			_ = buildSite()
 		}
 		pprof.WriteHeapProfile(f)
@@ -76,7 +76,7 @@ func benchmark(cmd *cobra.Command, args []string) error {
 		pprof.StartCPUProfile(f)
 		defer pprof.StopCPUProfile()
 		for i := 0; i < benchmarkTimes; i++ {
-			MainSite = nil
+			MainSites = nil
 			_ = buildSite()
 		}
 	}