This is currently slightly slower than master in a bench mark with 4 fairly big sites:

```
benchmark           old ns/op      new ns/op      delta
BenchmarkHugo-4     3146033049     3225819889     +2.54%

benchmark           old allocs     new allocs     delta
BenchmarkHugo-4     10557704       10850988       +2.78%

benchmark           old bytes      new bytes      delta
BenchmarkHugo-4     1628022920     1632976488     +0.30%
```

I guess this is mostly about doing "more work". I have done some general tuning in another PR (see #2378) after some profile work.



* [x] Add `Translations` and `AllTranslations` to Node.
* [x] Rework the site build process. Looping every site/language and doing the same for all of them kind of works, but isn't very effective nor elegant. Needs to extract the common tasks and do it once only.
* [x] Rework the documentation.
* [x] Fix all the `TODO(bep) multilingo` and  `TODO(bep) ml` comments.
* [x] Check the relevant URL funcs, `absURL` etc.
* [x] Add more tests
* [x]  Write redirect page to default lang on / when multilingual
* [x]  Write a global sitemap.xml
* [x]  Check that taxonomies can be configured on a per language basis
* [x] Add YAML config test  
* [x]  Check sitemap per language
* [x]  Check list command etc.
* [x] Move shortcode handling to "pre render". The site isn't really ready at the current point where the shortcodes gets rendered, which comes as a surprise to many. This is even more true now with the multilanguage feature (Site.Pages etc. will be empty). So we need to move this handling to a later stage.
* [x]  Watch: Handling adding and removing languages
* [x]  Check Blackfriday config per site
* [x]  Livereload inject
* [x] Check ref, relref  
* [x] implement https://github.com/spf13/hugo/issues/2312
* [x] Improve the default i18n vs missing bundles and translations 
* [x] See https://github.com/spf13/hugo/pull/2303/files#diff-d7b04948857565053b55f353ccb2bab2R296 while this is a temporary workaround until we get a proper registry for "all nodes and pages", this may be confusing to users expecting the home object in the translations object to be a full blown *home*. Now even the title is kind of wrong when it is set per language. We should fix this if it isn't too hard. A temporary "node map" maybe.
* [x] check data in Page.Site.Data from shortcodes + add some data to the ML tests
* [x] fix variables docs section -- there are some old and not valid info there now 
* [x] Double check all "reuse of nodes for RSS etc"
* [x] Double check the collection transfers
* [x] Check i18n template func in shortcode 
* [x]  Check shortcode changes vs livereload
* [x]  Fix Go 1.7 vet issues