@the1900 2016-07-27
I really delightful about this feature. If you don't mind I think taxonomy's should be multilingual too.

When see the taxonomy the url is fixed to `domain/taxonomy` but if url were `domain/lang/taxonomy` then it is more flexible I guess.