@CLAassistant 2016-08-05
[![CLA assistant check](https://cla-assistant.io/pull/badge/not_signed)](https://cla-assistant.io/spf13/hugo?pullRequest=2303) <br/>Thank you for your submission, we really appreciate it. Like many open source projects, we ask that you sign our [Contributor License Agreement](https://cla-assistant.io/spf13/hugo?pullRequest=2303) before we can accept your contribution.<br/>