@bep 2016-08-05
/cc @hacdias you might want to have a look at this re. your caddy plugin. Shout if you don't understand or if you need something special ...