@bep 2016-08-08
OK, this should work. A little pragmatic in some areas, but it is well tested.

As to how it works, see https://github.com/spf13/hugo/blob/66971e7e299facd61fb0e2e037ed290143dbc46e/docs/content/content/multilingual.md

Happy code review!

/cc @spf13