@bep 2016-08-09
Note to self, some weird recursive things happen if you get the TOML mapping wrong like this (see the blackfriday config). Probably related to the config watcher:

```
[Languages.en]
weight = 1

[Languages]
[Languages.nn]
weight = 2
title = "bep.is (Nynorsk)"
[Languages.en.blackfriday]
angledQuotes = true

[Languages]
[Languages.nn]
weight = 3
title = "bep.is (Nynorsk)"
[Languages.en.blackfriday]
angledQuotes = true
```