@gntech 2016-08-09
There is some problem reading the Language config when using a yaml-config file.
```
---
baseurl: "http://replace-this-with-your-hugo-site.com/"
languageCode: "en-us"
title: "My New Hugo Site"

Languages:
  en:
    weight: 1
  fr:
    weight: 2
```
```
Error: Error building site: Failed to parse multilingual config: Language config is not a map: map[]
```
When I try the same config in a toml-format it works as it should.