@gntech 2016-08-09
I am unable to use my data-files from shortcodes. This was possible before through ```.Page.Site.Data.``` Is this a bug or should I use some other variable now with the multilingual feature?