@bep 2016-08-10
@gntech see https://github.com/spf13/hugo/commit/85345d92df927fc1240343dca0b2502f65594374 I cannot reproduce it. I would rather say that this was an issue pre this branch. Now we evaluate the shortcodes as late as possible.