@bep 2016-08-10
To see what is needed to update a site to multilingual, see this commit from one of my sites:

https://github.com/bep/bep.is/commit/4d027b86b8aa8dfad0f7b4ed7b1baaf6b71c5616

It is a one-pager, so maybe not very typical, but you get the drift.

See it in action here:

http://bep.is/