@nils-holger 2016-08-10
Config.yaml doesn't work for me: `Error: Error building site: Failed to parse multilingual config: Language config is not a map: map[]`. I've been using the exact code as in the docs (except the toml issue of @gntech above). Config.toml works great.