@bep 2016-08-10
@nils-holger ok, I'll add a test for a yaml config, I'll admint I haven't tested that, but thought that should just work.