@bep 2016-08-10
@nils-holger see https://github.com/spf13/hugo/pull/2303/commits/663f09d949b21d3200453dfb1c84ee5c599ff91c re YAML not working. Thanks for the report. I was of course too lazy to test JSON ...