@mitchchn 2016-08-11
Apologies if this isn't the right PR to report this, but I've found a bug: partials do not have access to `$.Site.Data` when rendering translated pages (it's set to a `nil` map).