@bep 2016-08-11
@mitchchn I have tested that ... Is it something special about your case? Is the data in your theme? What OS?