@gntech 2016-08-11
@bep commit 3b4607f fixed the problem I had with reaching my data from shortcodes!