@mitchchn 2016-08-11
@bep That fixed the Data, thanks. There still seem to be things missing from .Site in translated content. $.Site.Taxonomies only includes translated pages for instance.