@bep 2016-08-11
That is by design. Taxonomies are site local, but they can have translations. I will double check others.