@gntech 2016-08-11
I have found a peculiar bug with the Permalink to translated content when I am on a node page. I have Swedish and English on my site with Swedish being the default language. When I am on the Swedish node page the link to the corresponding english node page works as expected however when I am on the English nodepage the Permalink back to Swedish nodepage points to the RSS "index.xml" of that node. A bit hard to explain but Very peculiar.

To summarize: The permalink to the translated counterpart of a node goes to the RSS page of the translated node.