@bep 2016-08-11
@gntech I understand that issue, but I thought I had fixed every places. Earlier the node instances from, say the home page and the home page's rss feed, where reused. Now when we use the node pointers to create a "translation map", this didn't work. It's an easy fix and thanks for testing.