@mitchchn 2016-08-12
One more: the `i18n` function does not appear to work in shortcodes.

```
ERROR: 2016/08/12 15:54:19 shortcode.go:590: error processing shortcode shortcodes/callout.html 
 ERR: template: shortcodes/callout.html:9:5: executing "shortcodes/callout.html" at <i18n "help">: error calling i18n: i18n not initialized, have you configured everything properly?
```

The same function call works fine in layouts.