@bep 2016-08-13
@mitchchn @gntech and others: My last commit fixes some issues when running in server watch mode, with reloading of changed shortcodes and similar. It would be helpful if you could take it for a spin.

I consider this branch solid enough now. @spf13 may have some to say, so I'll assign this issue to him.