@gntech 2016-08-14
@bep I have tested the latest code now and I have found no issues. Kjempebra! :)