@nils-holger 2016-08-15
@bep I've tested all documented features and all is working great now. Thanks for all time and efforts.