@bep 2016-08-16
For some performance numbers, see https://github.com/bep/hugo-benchmark