@gour 2016-08-28
I've built 0.17multilingual and after reading the docs tried to run:

`hugo --i18n-warnings | grep i18n
i18n|MISSING_TRANSLATION|en|wordCount`

but it reports about:

`Error: unknown flag: --i18n-warnings`

What do I miss?