@yitzhakbg 2016-08-31
### [Multipass build](https://heartbt.co.il/how/Multi-Language/)
simplifies multilingual blogs under one roof without having to invent wheels.