@yitzhakbg 2016-08-31
### [Multipass build](https://heartbt.co.il/how/Multi-Language/)
enables building each language site on its own, yet have them all under a single roof. Simplifies multilingual. No need to invent new wheels. The idea in short is to work with each content branch as if it were its own blog, unconcerned about other languages and merge the at build time. Layout, themes, asset branches etc. can be common to all.