@yitzhakbg 2016-08-31
Pardon me. Thanks for the correction

-- 
Yitzhak Bar Geva <https://www.linkedin.com/in/yitzhakbg>

On 1 September 2016 at 00:56:27, Bjørn Erik Pedersen (
notifications@github.com) wrote:

@yitzhakbg <https://github.com/yitzhakbg> please use the forum for
questions and marketing.

—
You are receiving this because you were mentioned.
Reply to this email directly, view it on GitHub
<https://github.com/spf13/hugo/pull/2303#issuecomment-243915090>, or mute
the thread
<https://github.com/notifications/unsubscribe-auth/AAA49IXiI55Cjn7ShuTZIoR3z7gOoUHMks5qlfh2gaJpZM4JVIOD>
.